requirements.txt
python-dotenv
SQLAlchemy
psycopg[binary]
uvicorn[standard]
fastapi[standard]
google-genai
Pillow
httpx
tqdm
